requirements.txt
django
djangorestframework
django-cors-headers
crewai
crewai-tools

